refactor(#991): align prompt-gen workflows with frame/shot model + thread frameId
Rename the two prompt-generation pipelines to the `-batch` parent + bare child
convention (matching motion-batch-workflow/motion-workflow), so names reflect
the data model (image prompt → frame_prompt_versions; motion prompt →
shot_prompt_versions) and the real fan-out unit:
- visual-prompt-workflow        → frame-prompt-batch-workflow  (FramePromptBatchWorkflow)
- visual-prompt-scene-workflow  → frame-prompt-workflow        (FramePromptWorkflow)
- motion-prompt-workflow        → motion-prompt-batch-workflow (MotionPromptBatchWorkflow)
- motion-prompt-scene-workflow  → motion-prompt-workflow       (MotionPromptWorkflow)

Updated bindings, wrangler.jsonc (all blocks), worker-configuration.d.ts,
server.ts re-exports, trigger-bindings paths, and the workflow input/result
types. Pure rename — durable step names, the `updateType` artifact strings, and
the VisualPrompt/MotionPrompt prompt-DATA types are deliberately untouched
(replay/contract identities, not workflow names).

Thread frameId (#991 no-DB-reads-in-workflows): the frame-prompt batch parent
resolves each shot's anchor frame id ONCE up front and passes it to the per-shot
children, so the child no longer calls getAnchorByShot. The anchor frame's
identity is immutable, so the single batched parent read is safe (not the racy
mutable-state category).

Gates: typecheck 0, wiring-consistency 17/17, 362 workflow/fn tests, lint 0
errors, knip clean.
