+++ src/lib/workflow/trigger-bindings.ts
@@ -32,8 +32,8 @@ const TRIGGER_TO_BINDING: Record<string, keyof CloudflareEnv> = {
   'library-location-sheet': 'LIBRARY_LOCATION_SHEET_WORKFLOW',
   'variant-image': 'SHOT_VARIANT_WORKFLOW',
   'upscale-variant': 'UPSCALE_SHOT_VARIANT_WORKFLOW',
-  'visual-prompt-scene': 'VISUAL_PROMPT_SCENE_WORKFLOW',
-  'motion-prompt-scene': 'MOTION_PROMPT_SCENE_WORKFLOW',
+  'frame-prompt': 'FRAME_PROMPT_WORKFLOW',
+  'motion-prompt': 'MOTION_PROMPT_WORKFLOW',
   'music-prompt': 'MUSIC_PROMPT_WORKFLOW',
   'recast-character': 'RECAST_CHARACTER_WORKFLOW',
   'location-matching': 'LOCATION_MATCHING_WORKFLOW',
@@ -41,8 +41,8 @@ const TRIGGER_TO_BINDING: Record<string, keyof CloudflareEnv> = {
   'talent-matching': 'TALENT_MATCHING_WORKFLOW',
   'character-sheet-from-bible': 'CHARACTER_BIBLE_WORKFLOW',
   'location-sheet-from-bible': 'LOCATION_BIBLE_WORKFLOW',
-  'visual-prompts': 'VISUAL_PROMPT_WORKFLOW',
-  'motion-prompts': 'MOTION_PROMPT_WORKFLOW',
+  'frame-prompts-batch': 'FRAME_PROMPT_BATCH_WORKFLOW',
+  'motion-prompts-batch': 'MOTION_PROMPT_BATCH_WORKFLOW',
   'motion-music-prompts': 'MOTION_MUSIC_PROMPTS_WORKFLOW',
   'regenerate-shots': 'REGENERATE_SHOTS_WORKFLOW',
   'recast-location': 'RECAST_LOCATION_WORKFLOW',
