+++ src/server.ts
@@ -66,8 +66,8 @@ export { LibraryTalentSheetWorkflow } from '@/lib/workflows/library-talent-sheet
 export { LibraryLocationSheetWorkflow } from '@/lib/workflows/library-location-sheet-workflow';
 export { ShotVariantWorkflow } from '@/lib/workflows/shot-variant-workflow';
 export { UpscaleShotVariantWorkflow } from '@/lib/workflows/upscale-shot-variant-workflow';
-export { VisualPromptSceneWorkflow } from '@/lib/workflows/visual-prompt-scene-workflow';
-export { MotionPromptSceneWorkflow } from '@/lib/workflows/motion-prompt-scene-workflow';
+export { FramePromptWorkflow } from '@/lib/workflows/frame-prompt-workflow';
+export { MotionPromptWorkflow } from '@/lib/workflows/motion-prompt-workflow';
 export { MusicPromptWorkflow } from '@/lib/workflows/music-prompt-workflow';
 export { RecastCharacterWorkflow } from '@/lib/workflows/recast-character-workflow';
 export { LocationMatchingWorkflow } from '@/lib/workflows/location-matching-workflow';
@@ -75,8 +75,8 @@ export { ShotImagesWorkflow } from '@/lib/workflows/shot-images-workflow';
 export { TalentMatchingWorkflow } from '@/lib/workflows/talent-matching-workflow';
 export { CharacterBibleWorkflow } from '@/lib/workflows/character-bible-workflow';
 export { LocationBibleWorkflow } from '@/lib/workflows/location-bible-workflow';
-export { VisualPromptWorkflow } from '@/lib/workflows/visual-prompt-workflow';
-export { MotionPromptWorkflow } from '@/lib/workflows/motion-prompt-workflow';
+export { FramePromptBatchWorkflow } from '@/lib/workflows/frame-prompt-batch-workflow';
+export { MotionPromptBatchWorkflow } from '@/lib/workflows/motion-prompt-batch-workflow';
 export { MotionMusicPromptsWorkflow } from '@/lib/workflows/motion-music-prompts-workflow';
 export { RegenerateShotsWorkflow } from '@/lib/workflows/regenerate-shots-workflow';
 export { RecastLocationWorkflow } from '@/lib/workflows/recast-location-workflow';
