+++ worker-configuration.d.ts
@@ -1,5 +1,5 @@
 /* eslint-disable */
-// Generated by Wrangler by running `wrangler types` (hash: 84f1b04760a91d8843246961a688aff3)
+// Generated by Wrangler by running `wrangler types` (hash: ce591809672c85c1371ca45e8b81299e)
 // Runtime types generated with workerd@1.20260511.1 2025-11-18 nodejs_compat
 declare namespace Cloudflare {
 	interface GlobalProps {
@@ -52,8 +52,8 @@ declare namespace Cloudflare {
 		LIBRARY_LOCATION_SHEET_WORKFLOW: Workflow<Parameters<import("./src/server").LibraryLocationSheetWorkflow['run']>[0]['payload']>;
 		SHOT_VARIANT_WORKFLOW: Workflow<Parameters<import("./src/server").ShotVariantWorkflow['run']>[0]['payload']>;
 		UPSCALE_SHOT_VARIANT_WORKFLOW: Workflow<Parameters<import("./src/server").UpscaleShotVariantWorkflow['run']>[0]['payload']>;
-		VISUAL_PROMPT_SCENE_WORKFLOW: Workflow<Parameters<import("./src/server").VisualPromptSceneWorkflow['run']>[0]['payload']>;
-		MOTION_PROMPT_SCENE_WORKFLOW: Workflow<Parameters<import("./src/server").MotionPromptSceneWorkflow['run']>[0]['payload']>;
+		FRAME_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").FramePromptWorkflow['run']>[0]['payload']>;
+		MOTION_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").MotionPromptWorkflow['run']>[0]['payload']>;
 		MUSIC_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").MusicPromptWorkflow['run']>[0]['payload']>;
 		RECAST_CHARACTER_WORKFLOW: Workflow<Parameters<import("./src/server").RecastCharacterWorkflow['run']>[0]['payload']>;
 		LOCATION_MATCHING_WORKFLOW: Workflow<Parameters<import("./src/server").LocationMatchingWorkflow['run']>[0]['payload']>;
@@ -61,8 +61,8 @@ declare namespace Cloudflare {
 		TALENT_MATCHING_WORKFLOW: Workflow<Parameters<import("./src/server").TalentMatchingWorkflow['run']>[0]['payload']>;
 		CHARACTER_BIBLE_WORKFLOW: Workflow<Parameters<import("./src/server").CharacterBibleWorkflow['run']>[0]['payload']>;
 		LOCATION_BIBLE_WORKFLOW: Workflow<Parameters<import("./src/server").LocationBibleWorkflow['run']>[0]['payload']>;
-		VISUAL_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").VisualPromptWorkflow['run']>[0]['payload']>;
-		MOTION_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").MotionPromptWorkflow['run']>[0]['payload']>;
+		FRAME_PROMPT_BATCH_WORKFLOW: Workflow<Parameters<import("./src/server").FramePromptBatchWorkflow['run']>[0]['payload']>;
+		MOTION_PROMPT_BATCH_WORKFLOW: Workflow<Parameters<import("./src/server").MotionPromptBatchWorkflow['run']>[0]['payload']>;
 		MOTION_MUSIC_PROMPTS_WORKFLOW: Workflow<Parameters<import("./src/server").MotionMusicPromptsWorkflow['run']>[0]['payload']>;
 		REGENERATE_SHOTS_WORKFLOW: Workflow<Parameters<import("./src/server").RegenerateShotsWorkflow['run']>[0]['payload']>;
 		RECAST_LOCATION_WORKFLOW: Workflow<Parameters<import("./src/server").RecastLocationWorkflow['run']>[0]['payload']>;
@@ -118,8 +118,8 @@ declare namespace Cloudflare {
 		LIBRARY_LOCATION_SHEET_WORKFLOW: Workflow<Parameters<import("./src/server").LibraryLocationSheetWorkflow['run']>[0]['payload']>;
 		SHOT_VARIANT_WORKFLOW: Workflow<Parameters<import("./src/server").ShotVariantWorkflow['run']>[0]['payload']>;
 		UPSCALE_SHOT_VARIANT_WORKFLOW: Workflow<Parameters<import("./src/server").UpscaleShotVariantWorkflow['run']>[0]['payload']>;
-		VISUAL_PROMPT_SCENE_WORKFLOW: Workflow<Parameters<import("./src/server").VisualPromptSceneWorkflow['run']>[0]['payload']>;
-		MOTION_PROMPT_SCENE_WORKFLOW: Workflow<Parameters<import("./src/server").MotionPromptSceneWorkflow['run']>[0]['payload']>;
+		FRAME_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").FramePromptWorkflow['run']>[0]['payload']>;
+		MOTION_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").MotionPromptWorkflow['run']>[0]['payload']>;
 		MUSIC_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").MusicPromptWorkflow['run']>[0]['payload']>;
 		RECAST_CHARACTER_WORKFLOW: Workflow<Parameters<import("./src/server").RecastCharacterWorkflow['run']>[0]['payload']>;
 		LOCATION_MATCHING_WORKFLOW: Workflow<Parameters<import("./src/server").LocationMatchingWorkflow['run']>[0]['payload']>;
@@ -127,8 +127,8 @@ declare namespace Cloudflare {
 		TALENT_MATCHING_WORKFLOW: Workflow<Parameters<import("./src/server").TalentMatchingWorkflow['run']>[0]['payload']>;
 		CHARACTER_BIBLE_WORKFLOW: Workflow<Parameters<import("./src/server").CharacterBibleWorkflow['run']>[0]['payload']>;
 		LOCATION_BIBLE_WORKFLOW: Workflow<Parameters<import("./src/server").LocationBibleWorkflow['run']>[0]['payload']>;
-		VISUAL_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").VisualPromptWorkflow['run']>[0]['payload']>;
-		MOTION_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").MotionPromptWorkflow['run']>[0]['payload']>;
+		FRAME_PROMPT_BATCH_WORKFLOW: Workflow<Parameters<import("./src/server").FramePromptBatchWorkflow['run']>[0]['payload']>;
+		MOTION_PROMPT_BATCH_WORKFLOW: Workflow<Parameters<import("./src/server").MotionPromptBatchWorkflow['run']>[0]['payload']>;
 		MOTION_MUSIC_PROMPTS_WORKFLOW: Workflow<Parameters<import("./src/server").MotionMusicPromptsWorkflow['run']>[0]['payload']>;
 		REGENERATE_SHOTS_WORKFLOW: Workflow<Parameters<import("./src/server").RegenerateShotsWorkflow['run']>[0]['payload']>;
 		RECAST_LOCATION_WORKFLOW: Workflow<Parameters<import("./src/server").RecastLocationWorkflow['run']>[0]['payload']>;
@@ -184,8 +184,8 @@ declare namespace Cloudflare {
 		LIBRARY_LOCATION_SHEET_WORKFLOW: Workflow<Parameters<import("./src/server").LibraryLocationSheetWorkflow['run']>[0]['payload']>;
 		SHOT_VARIANT_WORKFLOW: Workflow<Parameters<import("./src/server").ShotVariantWorkflow['run']>[0]['payload']>;
 		UPSCALE_SHOT_VARIANT_WORKFLOW: Workflow<Parameters<import("./src/server").UpscaleShotVariantWorkflow['run']>[0]['payload']>;
-		VISUAL_PROMPT_SCENE_WORKFLOW: Workflow<Parameters<import("./src/server").VisualPromptSceneWorkflow['run']>[0]['payload']>;
-		MOTION_PROMPT_SCENE_WORKFLOW: Workflow<Parameters<import("./src/server").MotionPromptSceneWorkflow['run']>[0]['payload']>;
+		FRAME_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").FramePromptWorkflow['run']>[0]['payload']>;
+		MOTION_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").MotionPromptWorkflow['run']>[0]['payload']>;
 		MUSIC_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").MusicPromptWorkflow['run']>[0]['payload']>;
 		RECAST_CHARACTER_WORKFLOW: Workflow<Parameters<import("./src/server").RecastCharacterWorkflow['run']>[0]['payload']>;
 		LOCATION_MATCHING_WORKFLOW: Workflow<Parameters<import("./src/server").LocationMatchingWorkflow['run']>[0]['payload']>;
@@ -193,8 +193,8 @@ declare namespace Cloudflare {
 		TALENT_MATCHING_WORKFLOW: Workflow<Parameters<import("./src/server").TalentMatchingWorkflow['run']>[0]['payload']>;
 		CHARACTER_BIBLE_WORKFLOW: Workflow<Parameters<import("./src/server").CharacterBibleWorkflow['run']>[0]['payload']>;
 		LOCATION_BIBLE_WORKFLOW: Workflow<Parameters<import("./src/server").LocationBibleWorkflow['run']>[0]['payload']>;
-		VISUAL_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").VisualPromptWorkflow['run']>[0]['payload']>;
-		MOTION_PROMPT_WORKFLOW: Workflow<Parameters<import("./src/server").MotionPromptWorkflow['run']>[0]['payload']>;
+		FRAME_PROMPT_BATCH_WORKFLOW: Workflow<Parameters<import("./src/server").FramePromptBatchWorkflow['run']>[0]['payload']>;
+		MOTION_PROMPT_BATCH_WORKFLOW: Workflow<Parameters<import("./src/server").MotionPromptBatchWorkflow['run']>[0]['payload']>;
 		MOTION_MUSIC_PROMPTS_WORKFLOW: Workflow<Parameters<import("./src/server").MotionMusicPromptsWorkflow['run']>[0]['payload']>;
 		REGENERATE_SHOTS_WORKFLOW: Workflow<Parameters<import("./src/server").RegenerateShotsWorkflow['run']>[0]['payload']>;
 		RECAST_LOCATION_WORKFLOW: Workflow<Parameters<import("./src/server").RecastLocationWorkflow['run']>[0]['payload']>;
