+++ wrangler.jsonc
@@ -165,14 +165,14 @@
       "class_name": "UpscaleShotVariantWorkflow",
     },
     {
-      "name": "visual-prompt-scene-workflow",
-      "binding": "VISUAL_PROMPT_SCENE_WORKFLOW",
-      "class_name": "VisualPromptSceneWorkflow",
+      "name": "frame-prompt-workflow",
+      "binding": "FRAME_PROMPT_WORKFLOW",
+      "class_name": "FramePromptWorkflow",
     },
     {
-      "name": "motion-prompt-scene-workflow",
-      "binding": "MOTION_PROMPT_SCENE_WORKFLOW",
-      "class_name": "MotionPromptSceneWorkflow",
+      "name": "motion-prompt-workflow",
+      "binding": "MOTION_PROMPT_WORKFLOW",
+      "class_name": "MotionPromptWorkflow",
     },
     {
       "name": "music-prompt-workflow",
@@ -210,14 +210,14 @@
       "class_name": "LocationBibleWorkflow",
     },
     {
-      "name": "visual-prompt-workflow",
-      "binding": "VISUAL_PROMPT_WORKFLOW",
-      "class_name": "VisualPromptWorkflow",
+      "name": "frame-prompt-batch-workflow",
+      "binding": "FRAME_PROMPT_BATCH_WORKFLOW",
+      "class_name": "FramePromptBatchWorkflow",
     },
     {
-      "name": "motion-prompt-workflow",
-      "binding": "MOTION_PROMPT_WORKFLOW",
-      "class_name": "MotionPromptWorkflow",
+      "name": "motion-prompt-batch-workflow",
+      "binding": "MOTION_PROMPT_BATCH_WORKFLOW",
+      "class_name": "MotionPromptBatchWorkflow",
     },
     {
       "name": "motion-music-prompts-workflow",
@@ -397,14 +397,14 @@
           "class_name": "UpscaleShotVariantWorkflow",
         },
         {
-          "name": "visual-prompt-scene-workflow",
-          "binding": "VISUAL_PROMPT_SCENE_WORKFLOW",
-          "class_name": "VisualPromptSceneWorkflow",
+          "name": "frame-prompt-workflow",
+          "binding": "FRAME_PROMPT_WORKFLOW",
+          "class_name": "FramePromptWorkflow",
         },
         {
-          "name": "motion-prompt-scene-workflow",
-          "binding": "MOTION_PROMPT_SCENE_WORKFLOW",
-          "class_name": "MotionPromptSceneWorkflow",
+          "name": "motion-prompt-workflow",
+          "binding": "MOTION_PROMPT_WORKFLOW",
+          "class_name": "MotionPromptWorkflow",
         },
         {
           "name": "music-prompt-workflow",
@@ -442,14 +442,14 @@
           "class_name": "LocationBibleWorkflow",
         },
         {
-          "name": "visual-prompt-workflow",
-          "binding": "VISUAL_PROMPT_WORKFLOW",
-          "class_name": "VisualPromptWorkflow",
+          "name": "frame-prompt-batch-workflow",
+          "binding": "FRAME_PROMPT_BATCH_WORKFLOW",
+          "class_name": "FramePromptBatchWorkflow",
         },
         {
-          "name": "motion-prompt-workflow",
-          "binding": "MOTION_PROMPT_WORKFLOW",
-          "class_name": "MotionPromptWorkflow",
+          "name": "motion-prompt-batch-workflow",
+          "binding": "MOTION_PROMPT_BATCH_WORKFLOW",
+          "class_name": "MotionPromptBatchWorkflow",
         },
         {
           "name": "motion-music-prompts-workflow",
@@ -669,14 +669,14 @@
           "class_name": "UpscaleShotVariantWorkflow",
         },
         {
-          "name": "visual-prompt-scene-workflow",
-          "binding": "VISUAL_PROMPT_SCENE_WORKFLOW",
-          "class_name": "VisualPromptSceneWorkflow",
+          "name": "frame-prompt-workflow",
+          "binding": "FRAME_PROMPT_WORKFLOW",
+          "class_name": "FramePromptWorkflow",
         },
         {
-          "name": "motion-prompt-scene-workflow",
-          "binding": "MOTION_PROMPT_SCENE_WORKFLOW",
-          "class_name": "MotionPromptSceneWorkflow",
+          "name": "motion-prompt-workflow",
+          "binding": "MOTION_PROMPT_WORKFLOW",
+          "class_name": "MotionPromptWorkflow",
         },
         {
           "name": "music-prompt-workflow",
@@ -714,14 +714,14 @@
           "class_name": "LocationBibleWorkflow",
         },
         {
-          "name": "visual-prompt-workflow",
-          "binding": "VISUAL_PROMPT_WORKFLOW",
-          "class_name": "VisualPromptWorkflow",
+          "name": "frame-prompt-batch-workflow",
+          "binding": "FRAME_PROMPT_BATCH_WORKFLOW",
+          "class_name": "FramePromptBatchWorkflow",
         },
         {
-          "name": "motion-prompt-workflow",
-          "binding": "MOTION_PROMPT_WORKFLOW",
-          "class_name": "MotionPromptWorkflow",
+          "name": "motion-prompt-batch-workflow",
+          "binding": "MOTION_PROMPT_BATCH_WORKFLOW",
+          "class_name": "MotionPromptBatchWorkflow",
         },
         {
           "name": "motion-music-prompts-workflow",
